[feat] sperate jpeg.rs
